ML-Production/requirements.txt
torch
torchvision
flask
numpy
Pillow